server/src/config.py
import os
import environ
import psycopg2
from flask_sqlalchemy import SQLAlchemy


class Config(object):
    DEBUG = False
    TESTING = False
    CSRF_ENABLED = True
    SECRET_KEY = env("SECRET_KEY")
    SQLALCHEMY_DATABASE_URI = env('DB_URL')


class ProductionConfig(Config):
    DEBUG = False


class StagingConfig(Config):
    DEVELOPMENT = True
    DEBUG = True


class DevelopmentConfig(Config):
    DEVELOPMENT = True
    DEBUG = True


class TestingConfig(Config):
    TESTING = True
