server/requirements.txt
alembic==1.11.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.5
environ==1.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
Werkzeug==2.3.6
